python_flask/requirements.txt
numpy
pandas
Flask
waitress
mysql.connector
databases[mysql]